Added support to run qsfix directly on TiVo files (bypassing decrypt)
git-svn-id: https://svn.code.sf.net/p/kmttg/code/trunk@754 b0df3fdd-2057-4f73-b226-dfc9190e8f7a

+++ src/com/tivo/kmttg/task/qsfix.java
@@ -24,6 +24,7 @@ public class qsfix {
    String  vrdscript = null;
    String  vrdscript_temp = null;
    String  cscript = null;
+   String  sourceFile = null;
    private backgroundProcess process;
    private jobData job;
 
@@ -54,12 +55,18 @@ public class qsfix {
          log.error("File does not exist: " + cscript);
          schedule = false;
       }
+      
+      if (config.VrdDecrypt == 0) {
+         sourceFile = job.mpegFile;
+      } else {
+         sourceFile = job.tivoFile;
+      }
                   
-      if ( ! file.isFile(job.mpegFile) ) {
-         log.error("mpeg file not found: " + job.mpegFile);
+      if ( ! file.isFile(sourceFile) ) {
+         log.error("source file not found: " + sourceFile);
          schedule = false;
       }
-      
+            
       if (schedule) {
          // Create sub-folders for output file if needed
          if ( ! jobMonitor.createSubFolders(job.mpegFile_fix, job) ) schedule = false;
@@ -68,9 +75,9 @@ public class qsfix {
       if (config.VrdQsFilter == 1 && schedule) {
          log.warn("VideoRedo video dimensions filter is enabled - making custom VRD script");
          // Want QSFix run with video dimension filter enabled
-         Hashtable<String,String> dimensions = ffmpegGetVideoDimensions(job.mpegFile);
+         Hashtable<String,String> dimensions = ffmpegGetVideoDimensions(sourceFile);
          if (dimensions == null) {
-            log.error("VRD QS Filter enabled but unable to determine video dimensions of file: " + job.mpegFile);
+            log.error("VRD QS Filter enabled but unable to determine video dimensions of file: " + sourceFile);
             schedule = false;
          } else {    
             log.warn("VideoRedo video dimensions filter set to: x=" + dimensions.get("x") + ", y=" + dimensions.get("y"));
@@ -104,14 +111,14 @@ public class qsfix {
       command.add(cscript);
       command.add("//nologo");
       command.add(vrdscript);
-      command.add(job.mpegFile);
+      command.add(sourceFile);
       command.add(job.mpegFile_fix);
       command.add("/t1");
       command.add("/d");
       command.add("/q");
       command.add("/na");
       process = new backgroundProcess();
-      log.print(">> Running qsfix on " + job.mpegFile + " ...");
+      log.print(">> Running qsfix on " + sourceFile + " ...");
       if ( process.run(command) ) {
          log.print(process.toString());
       } else {
@@ -145,7 +152,7 @@ public class qsfix {
                // Update status in job table
                String s = String.format("%.2f MB", (float)file.size(job.mpegFile_fix)/Math.pow(2,20));
                String t = jobMonitor.getElapsedTime(job.time);
-               int pct = Integer.parseInt(String.format("%d", file.size(job.mpegFile_fix)*100/file.size(job.mpegFile)));
+               int pct = Integer.parseInt(String.format("%d", file.size(job.mpegFile_fix)*100/file.size(sourceFile)));
                               
                if ( jobMonitor.isFirstJobInMonitor(job) ) {
                   // Update STATUS column 
@@ -189,11 +196,14 @@ public class qsfix {
             log.warn("qsfix job completed: " + jobMonitor.getElapsedTime(job.time));
             log.print("---DONE---");
             // Rename mpegFile_fix to mpegFile
-            Boolean result = file.delete(job.mpegFile);
-            if ( ! result ) {
-            	log.error("Failed to delete file in preparation for rename: " + job.mpegFile);
-            	if (vrdscript_temp != null) file.delete(vrdscript_temp);
-            	return false;
+            Boolean result;
+            if (config.VrdDecrypt == 0) {
+               result = file.delete(job.mpegFile);
+               if ( ! result ) {
+               	log.error("Failed to delete file in preparation for rename: " + job.mpegFile);
+               	if (vrdscript_temp != null) file.delete(vrdscript_temp);
+               	return false;
+               }
             }
             result = file.rename(job.mpegFile_fix, job.mpegFile);
             if (result)
